Reduce test_utils.mock_operators to only that which is reusued
There was a lot in tests.test_utils.mock_* that was only used by the
Hive provider -- all of that has been removed and put in
tests.providers.apache.hive

Anything else that was only used by a single test has been moved in to
that specific test

(The main driver for this was to remove the import of HiveOperator from
mock_operators to make it easier to run "core" tests without needing
_all_ of the extras installed.)
