+++ tests/providers/apache/hive/transfers/test_hive_to_samba.py
@@ -23,9 +23,25 @@ import pytest
 
 from airflow import PY39
 from airflow.providers.apache.hive.transfers.hive_to_samba import HiveToSambaOperator
+from airflow.providers.samba.hooks.samba import SambaHook
 from airflow.utils.operator_helpers import context_to_airflow_vars
-from tests.providers.apache.hive import DEFAULT_DATE, TestHiveEnvironment
-from tests.test_utils.mock_hooks import MockHiveServer2Hook, MockSambaHook
+from tests.providers.apache.hive import (
+    DEFAULT_DATE,
+    MockConnectionCursor,
+    MockHiveServer2Hook,
+    TestHiveEnvironment,
+)
+
+
+class MockSambaHook(SambaHook):
+    def __init__(self, *args, **kwargs):
+        self.conn = MockConnectionCursor()
+        self.conn.execute = MagicMock()
+        self.get_conn = MagicMock(return_value=self.conn)
+        super().__init__(*args, **kwargs)
+
+    def get_connection(self, *args):
+        return self.conn
 
 
 @pytest.mark.skipif(
